remove generate script, integrate to runtest.sh
generate_expected_output.sh and runtest.sh are pretty much the same,
just the output dir of the rpmvalidator.sh differs and no asserts in
generate_expected_output.sh. To lower the maintainance burden this
is now merged and generating the expected output can be done with
the option -g/--generate.

Signed-off-by: Reto Zingg <reto.zingg@jolla.com>