[script] check that rpath is set if lib is shipped
If own libraries are shipped with the rpm package, the rpath in the
binary must be set, otherwise these libraries are not taken into use by
the dynamic linker. This is specially a problem with libraries which are
otherwise available on the platform.

[mersdk@SailfishSDK tests]$ ./runtest.sh
testHarbourGoodArm
testHarbourGoodArmNoColor
testHarbourBadArm
testHarbourBadArmNoColor
testHarbourGoodIntel
testHarbourBadIntel
testBadFileName
testNotAnRpm
testBadFileNameNoColor
testNotAnRpmNoColor
testHarborBadArm

Ran 11 tests.

OK

Signed-off-by: Reto Zingg <reto.zingg@jolla.com>