Add PyOtherSide 1.4 & 1.5 to allowed QML imports
PyOtherSide has been updated to 1.5 in Sailfish OS quite long ago, but the allowed imports file has apparently not been updated. So add versions 1.4 & 1.5 of the PyOtherSide QML module to the allowed QML import list so that developers of Python/QML applications can make use of the new features and fixes added.